Leveling Bar
I put in a leveling bar, ( I understand that this cliche implementation might not be used). But hopefully the leveling math could. I didn't have anything to tie exp points to so as of right now the player will gain exp by pressing space. The progress bar named "Leveling_Progress" should update with the exp correctly and reset when a new level is achieved.